+++ SalesPipeline.Utils/Resources/PreApprove/PayScheduleFactor.cs
@@ -8,7 +8,7 @@ namespace SalesPipeline.Utils.Resources.PreApprove
 {
 	public class PayScheduleFactor
 	{
-		public decimal lookupValue { get; set; }
+		public double lookupValue { get; set; }
 		/// <summary>
 		/// สินเชื่อ
 		/// </summary>
